Rename the default controller.
