+++ src/App/Routes.php
@@ -2,5 +2,5 @@
 
 declare(strict_types=1);
 
-$app->get('/', 'App\Controller\Base\BaseController:getHelp');
-$app->get('/status', 'App\Controller\Base\BaseController:getStatus');
+$app->get('/', 'App\Controller\Home:getHelp');
+$app->get('/status', 'App\Controller\Home:getStatus');
